requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-core==2.16.3
pytest==8.1.1
python-multipart==0.0.9
pyyaml==6.0.1
redis==5.0.3
ruff==0.3.5
setuptools==69.2.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing-extensions==4.11.0
uvicorn==0.29.0
virtualenv==20.25.1
